+++ src/rpc/mining.cpp
@@ -242,9 +242,12 @@ static UniValue getmininginfo(const Config &config,
             RPCResult{
                 "{\n"
                 "  \"blocks\": nnn,             (numeric) The current block\n"
-                "  \"currentblocksize\": nnn,   (numeric) The last block size\n"
-                "  \"currentblocktx\": nnn,     (numeric) The last block "
-                "transaction\n"
+                "  \"currentblocksize\": nnn,   (numeric, optional) The block "
+                "size of the last assembled block (only present if a block was "
+                "ever assembled)\n"
+                "  \"currentblocktx\": nnn,     (numeric, optional) The number "
+                "of block transactions of the last assembled block (only "
+                "present if a block was ever assembled)\n"
                 "  \"difficulty\": xxx.xxxxx    (numeric) The current "
                 "difficulty\n"
                 "  \"networkhashps\": nnn,      (numeric) The network hashes "
@@ -266,8 +269,12 @@ static UniValue getmininginfo(const Config &config,
 
     UniValue obj(UniValue::VOBJ);
     obj.pushKV("blocks", int(::ChainActive().Height()));
-    obj.pushKV("currentblocksize", uint64_t(nLastBlockSize));
-    obj.pushKV("currentblocktx", uint64_t(nLastBlockTx));
+    if (BlockAssembler::m_last_block_size) {
+        obj.pushKV("currentblocksize", *BlockAssembler::m_last_block_size);
+    }
+    if (BlockAssembler::m_last_block_num_txs) {
+        obj.pushKV("currentblocktx", *BlockAssembler::m_last_block_num_txs);
+    }
     obj.pushKV("difficulty", double(GetDifficulty(::ChainActive().Tip())));
     obj.pushKV("networkhashps", getnetworkhashps(config, request));
     obj.pushKV("pooledtx", uint64_t(g_mempool.size()));